Add files via upload
The Java program works as a simulator, the simulator helps the user to buy stocks that are available in the market and let them sell stocks as well if they already have some in their account. Initially, the user does not have any stock but moving forward this code will also save the user’s portfolio and save all the data. However, the simulator calculates all the trades that the user makes and stores them in their portfolio. Initially, the user has $20000 in their account with which they can make purchases. If the user is not sure about the current prices of the stock, the simulator will have an option with which the user can see the prices of the stock.
There are two Java files out of which one is for all the stocks, and one is the main program. Both have one constructer in them and then there are several methods that have been used while working on this simulator, and a couple of classes were used to get values and for calculating the total amount of the purchase. For example, a method for adding stock to an array list, and a method for buying stocks, this method also checks if the user has enough balance in their account, and a method for selling stocks if the user already has them, otherwise let them know that they don’t have enough stocks to sell. Moreover, there are a few more methods, for example, there is a method that keeps track of the portfolio value, and this is then called in the display portfolio which works as displaying the portfolio, this portfolio shows the user how much balance they are left with and shows all the stocks they have. To make it easy for users, I have added the prices here as well in the portfolio, so that users can see what rate each stock was when they bought it. And yes, there is a method used to display all the prices of the stock separately. Next up, I have also added a method that checks what the user wants to do and then prompts that thing accordingly. Other than this, there are a few more methods, that are getters and setters, so they are very small methods.
And then, finally, I have an action enum type in the program for users’ input, like display prices, display portfolio, buy or sell stocks, and lastly exit option. These are used in the switch selection for the user input and then will accordingly call the method.
This simulator is currently in the development process, so soon, we will see many options that the simulator will provide to make the trading of stock interesting and helpful. As mentioned above, I am also working on how this code can let users save their portfolios. This will make it more interesting, as they might want to continue the next day from where they left off.